# backend/fastApi_backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/get-meetings/{page}/{pagesize}": {
      "get": {
        "summary": "Get Meetings List",
        "operationId": "get_meetings_list_get_meetings__page___pagesize__get",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page"
            }
          },
          {
            "name": "pagesize",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListMeetings"
                }
              }
            }
          },
          "404": {
            "description": "No meetings found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meeting-details/{meeting_id}": {
      "get": {
        "summary": "Get Meeting Detials",
        "operationId": "get_meeting_detials_meeting_details__meeting_id__get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExistedMeeting"
                }
              }
            }
          },
          "404": {
            "description": "Meeting not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/get-people": {
      "get": {
        "summary": "Get People List",
        "operationId": "get_people_list_get_people_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Guest"
                  },
                  "type": "array",
                  "title": "Response Get People List Get People Get"
                }
              }
            }
          }
        }
      }
    },
    "/get-agendas": {
      "get": {
        "summary": "Get Agendas List",
        "operationId": "get_agendas_list_get_agendas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Agenda"
                  },
                  "type": "array",
                  "title": "Response Get Agendas List Get Agendas Get"
                }
              }
            }
          }
        }
      }
    },
    "/upload-files/": {
      "post": {
        "summary": "Create Upload Files",
        "operationId": "create_upload_files_upload_files__post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_upload_files_upload_files__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/new-meeting": {
      "post": {
        "summary": "Add Meeting",
        "operationId": "add_meeting_new_meeting_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseMeeting"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Meeting need at least one meeting address or online address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/update-meeting": {
      "put": {
        "summary": "Update Meeting",
        "operationId": "update_meeting_update_meeting_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExistedMeeting"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Meeting with id #{edited_meeting.id} not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/delete-meeting/{meeting_id}": {
      "delete": {
        "summary": "Delete Meetings",
        "operationId": "delete_meetings_delete_meeting__meeting_id__delete",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/get-project-info": {
      "get": {
        "summary": "Get Project Info",
        "operationId": "get_project_info_get_project_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDataExternal"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Agenda": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "order": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order"
          }
        },
        "type": "object",
        "title": "Agenda"
      },
      "BaseMeeting": {
        "properties": {
          "meeting_type": {
            "$ref": "#/components/schemas/MeetingType"
          },
          "meeting_name": {
            "type": "string",
            "minLength": 1,
            "title": "Meeting Name"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date"
          },
          "meeting_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Address"
          },
          "online_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Online Address"
          },
          "guests": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Guest"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guests"
          },
          "tasksList": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Task"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taskslist"
          },
          "agenda": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Agenda"
              },
              {
                "type": "null"
              }
            ]
          },
          "documents": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documents"
          }
        },
        "type": "object",
        "required": [
          "meeting_type",
          "meeting_name",
          "start_date",
          "end_date"
        ],
        "title": "BaseMeeting"
      },
      "Body_create_upload_files_upload_files__post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_create_upload_files_upload_files__post"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "ExistedMeeting": {
        "properties": {
          "meeting_type": {
            "$ref": "#/components/schemas/MeetingType"
          },
          "meeting_name": {
            "type": "string",
            "minLength": 1,
            "title": "Meeting Name"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date"
          },
          "meeting_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Address"
          },
          "online_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Online Address"
          },
          "guests": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Guest"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guests"
          },
          "tasksList": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Task"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taskslist"
          },
          "agenda": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Agenda"
              },
              {
                "type": "null"
              }
            ]
          },
          "documents": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documents"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "meeting_type",
          "meeting_name",
          "start_date",
          "end_date"
        ],
        "title": "ExistedMeeting"
      },
      "Guest": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "surname": {
            "type": "string",
            "title": "Surname"
          },
          "job_position": {
            "type": "string",
            "title": "Job Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "surname",
          "job_position"
        ],
        "title": "Guest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MeetingType": {
        "type": "string",
        "enum": [
          "boardMeeting",
          "generalAssembly",
          "other"
        ],
        "title": "MeetingType"
      },
      "PagedListMeetings": {
        "properties": {
          "meetings": {
            "items": {
              "$ref": "#/components/schemas/ShortMeeting"
            },
            "type": "array",
            "title": "Meetings"
          },
          "total_length": {
            "type": "integer",
            "title": "Total Length"
          }
        },
        "type": "object",
        "required": [
          "meetings",
          "total_length"
        ],
        "title": "PagedListMeetings"
      },
      "ProjectDataExternal": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "surname": {
            "type": "string",
            "title": "Surname"
          },
          "projectName": {
            "type": "string",
            "title": "Projectname"
          },
          "projectVersion": {
            "type": "string",
            "title": "Projectversion"
          },
          "indexNumber": {
            "type": "integer",
            "title": "Indexnumber"
          }
        },
        "type": "object",
        "required": [
          "name",
          "surname",
          "projectName",
          "projectVersion",
          "indexNumber"
        ],
        "title": "ProjectDataExternal"
      },
      "ShortMeeting": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "meeting_type": {
            "$ref": "#/components/schemas/MeetingType"
          },
          "meeting_name": {
            "type": "string",
            "minLength": 1,
            "title": "Meeting Name"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "meeting_type",
          "meeting_name",
          "start_date",
          "end_date"
        ],
        "title": "ShortMeeting"
      },
      "Task": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "Task"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}